chore: layout tweaks
